+++ build_runner_core/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Refactor `BuildCacheReader` to `BuildCacheAssetPathProvider`.
 - Refactor `FileBasedAssetReader` and `FileBasedAssetWriter` to `ReaderWriter`.
 - Move `BuildStepImpl` to `build_runner_core`, use `SingleStepReader` directly.
+- Remove `BuildCacheWriter`, functionality is handled by `AssetPathProvider`.
 
 ## 8.0.0